rename accordion tabs file to follow convention now that we don't need to hack alphabetism for dependency
