+++ docs/layout/default.html
@@ -109,7 +109,7 @@
   <script src="assets/js/plugins/foundation.tabs.js"></script>
   <script src="assets/js/plugins/foundation.toggler.js"></script>
   <script src="assets/js/plugins/foundation.tooltip.js"></script>
-  <script src="assets/js/plugins/foundation.zf.responsiveAccordionTabs.js"></script>
+  <script src="assets/js/plugins/foundation.responsiveAccordionTabs.js"></script>
 
   <script src="assets/js/docs.js"></script>
   <script type="text/javascript" src="https://intercom.zurb.com/scripts/zcom.js"></script>
