+++ js/entries/plugins/foundation.responsiveAccordionTabs.js
@@ -1,5 +1,5 @@
 import { Foundation } from './foundation.core';
 
-import { ResponsiveAccordionTabs } from '../../foundation.zf.responsiveAccordionTabs';
+import { ResponsiveAccordionTabs } from '../../foundation.responsiveAccordionTabs';
 Foundation.plugin(ResponsiveAccordionTabs, 'ResponsiveAccordionTabs');
