+++ js/foundation.responsiveAccordionTabs.js
